Update KMA_RNA_RAW_counts.csv
********VERY IMPORTANT CHANGE***********
This file has the raw K.marxianus data - during analysis two conditions were swapped for each other. Specifically this occurred during the differential expression analysis (the samples switched were osmotic and low pH). This is because the script LoadRNAdata.R expects low pH data to be in columns H-J, while in the previous version of this file (KMA_RNA_RAW_counts.csv, osmotic samples were in this location H-J.
Due to this realization, Ortholog Venn diagrams previously made for low pH and osmotic conditions will be incorrect. Downstream analysis will need to be adjusted to ensure that the labels are changed for all graphs, volcanoes, etc.