Re-ran the analysis for KMA - two samples were previously switched
DE was previously run from kma_RNA_RAW_counts. The script was written with the asumption that samples were Stan. hit,lowpH, then osmotic. Samples were Stan, HiT, Osm, and LpH for K.marx. This led to the DE for LpH and Osmotic to be switched.

Previously, I corrected the order of the samples on the .csv sheet. Here in Results2, I reran the analysis. The only change in the analysis is that scale.=FALSE for the PCA plot. Quick checks suggest that the datasets are in the correct order now.