Remapped YLI data using strand-specificity in featureCounts
Previously, STAR_feat2 was being used, which did not require strand specificity. This has been replaced by STAR_feat4, which was used to remake the yli_RNA_RAW_counts.csv file. This change removed the mapping of ~3.5% of total reads! In all, this removes ~600 genes from the RNAseq analysis due to having too few reads to pass the CPM threshold. Several other genes change, some that were previously significant are no longer significant...but overall, the accuracy of mapping should be increased. Also in this commit is a rerun of the pipeline (without changes) to update the YLI dataset.