Suggestions for Readme
Added some suggested changes to clarify for novices like myself. Also added some additional information at the top of the readme, which may help get the readers attention to focus on what this repo can do that is unique.

None of these changes are critical, so feel free to use whichever you like.