Changed the multiOmics_Analysis Main Script Name
Suggested new name for this script to bring it to the top of the directory and make it stand out more.

Also, added notes that spell out to the reader that Line 50 is the only thing they need to change for a basic analysis. Also, I changed the directory name in line 50 to reflect the users directory name "OrthOmics"