Added a multi-copy ortholog processing script
Beta version of preprocessOrthoResultsAnyCopy#.R script that identifies when a test organism(i.e. K.marxianus) shares an ortholog (regardless of copy number) with a set of query species. This analysis is designed to run on the directories created via orthofinder, and is not limited by the number of query organisms used.